DIRECTOR: High IDs for custom scrollable border
This patch bumps IDs of scrollable border to 0x1000 and above, so that
they don't conflict with IDs normal borders, additionally to support this
change windowType is changed from `byte` type to `uint32` type.

+++ graphics/macgui/macwindowmanager.h
@@ -359,8 +359,8 @@ public:
 	void cleanupDataBundle();
 	void cleanupDesktopBmp();
 
-	BorderOffsets getBorderOffsets(byte windowType);
-	Common::SeekableReadStream *getBorderFile(byte windowType, uint32 flags);
+	BorderOffsets getBorderOffsets(uint32 windowType);
+	Common::SeekableReadStream *getBorderFile(uint32 windowType, uint32 flags);
 	Common::SeekableReadStream *getFile(const Common::String &filename);
 
 	void setTextInClipboard(const Common::U32String &str);
